requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic

passlib[bcrypt]
python-jose[cryptography]
python-multipart
bcrypt

python-dotenv

stanza
nltk
pyspellchecker
pandas